server/requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
boto3==1.42.90
botocore==1.42.90
celery==5.6.3
celery-types==0.26.0
channels==4.3.2
channels_redis==4.3.0
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.13
django-cors-headers==4.9.0
django-redis==6.0.0
django-storages==1.14.6
django-stubs==6.0.2
django-stubs-ext==6.0.2
djangorestframework==3.17.1
djangorestframework-stubs==3.16.9
djangorestframework_simplejwt==5.5.1
jmespath==1.1.0
kombu==5.6.2
msgpack==1.1.2
packaging==26.1
pillow==12.2.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.4.0
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
types-PyYAML==6.0.12.20260408
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
